% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field-data-funs.R
\name{anem_dive}
\alias{anem_dive}
\title{Pull in anemones and the diveinfo that goes with them}
\usage{
anem_dive(anems_ids)
}
\arguments{
\item{anems_ids}{}
}
\value{
A tibble/data frame
}
\description{
Pull in anemones and the diveinfo that goes with them
}
\examples{
anem_info <- anem_dive(2001, 0134)
}
