% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field-data-funs.R
\name{date_range_dive}
\alias{date_range_dive}
\title{Pull all of the dives from a specific date range}
\usage{
date_range_dive(begin_date, end_date)
}
\arguments{
\item{begin_date}{A character date}

\item{end_date}{A character date}
}
\value{
A table of dives
}
\description{
Pull all of the dives from a specific date range
}
\examples{
dives <- date_range_dive("2017-01-01", "2017-12-30")
}
