% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field-data-funs.R
\name{fish_anem_dive}
\alias{fish_anem_dive}
\title{Pull a large joined table that includes fish, anemone, and diveinfo
This function pulls in clownfish data and joins all related anemone and diveinfo into a table in R.}
\usage{
fish_anem_dive()
}
\value{
A tibble/dataframe
}
\description{
Pull a large joined table that includes fish, anemone, and diveinfo
This function pulls in clownfish data and joins all related anemone and diveinfo into a table in R.
}
\examples{
fish_of_interest <- c("APCL18_201", "APCL17_353", "APCL15_013")
fish_info <- fish_anem_dive() \%>\%
filter(sample_id \%in\% fish_of_interest)
}
