% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field-data-funs.R
\name{get_anem}
\alias{get_anem}
\title{Pull the anemone table from the leyte database}
\usage{
get_anem()
}
\value{
A tibble/data frame
}
\description{
This function pulls in the entire anemone table which can be filtered for a more specific query.
}
\examples{
anem <- get_anem() \%>\%
filter(anem_obs == 12)
}
