% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field-data-funs.R
\name{get_dive}
\alias{get_dive}
\title{Pull the diveinfo table from the leyte database
This function pulls the entire diveinfo table which can be filtered for a more specific query}
\usage{
get_dive()
}
\value{
A tibble/data frame
}
\description{
Pull the diveinfo table from the leyte database
This function pulls the entire diveinfo table which can be filtered for a more specific query
}
\examples{
dive <- get_dive()

}
