% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field-data-funs.R
\name{get_fish}
\alias{get_fish}
\title{Pull the clownfish table from the leyte database}
\usage{
get_fish()
}
\value{
A tibble/data frame
}
\description{
This function pulls in the entire clownfish table which can be filtered for a more specific query.
}
\examples{
fish <- get_fish() \%>\%
filter(sample_id == "APCL18_201")
}
