% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab-data-funs.R
\name{plate_from_db}
\alias{plate_from_db}
\title{Maps the well locations of samples from the database}
\usage{
plate_from_db(table_name, ...)
}
\arguments{
\item{...}{A column identifier}
}
\value{
A table and a platemap
}
\description{
Maps the well locations of samples from the database
}
\examples{

extractions <- plate_from_db(extractions, extraction_id)
}
