% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-connections.R
\name{read_db}
\alias{read_db}
\title{Access the Pinsky Lab database}
\usage{
read_db(db_name)
}
\arguments{
\item{db_name}{name of a database}
}
\value{
database connection
}
\description{
Access the Pinsky Lab database
}
