% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field-data-funs.R
\name{sample_latlon}
\alias{sample_latlon}
\title{Return the lat lon for a set of sample_ids}
\usage{
sample_latlon(sample_ids)
}
\arguments{
\item{sample_ids}{}
}
\value{
A tibble/data frame that contains the columns sample_id, lat, and lon
}
\description{
Return the lat lon for a set of sample_ids
}
\examples{
location <- sample_latlon(c("APCL15_201", "APCL17_201"))
}
